Add AvailableLanguages wrapper struct
