+++ crates/language/src/language_registry.rs
@@ -1,7 +1,9 @@
+use crate::available_languages::AvailableLanguage;
 use crate::{
     CachedLspAdapter, File, Language, LanguageConfig, LanguageId, LanguageMatcher,
     LanguageServerName, LspAdapter, ManifestName, PLAIN_TEXT, ToolchainLister,
-    language_settings::all_language_settings, task_context::ContextProvider, with_parser,
+    available_languages::AvailableLanguages, language_settings::all_language_settings,
+    task_context::ContextProvider, with_parser,
 };
 use anyhow::{Context as _, Result, anyhow};
 use collections::{FxHashMap, HashMap, HashSet, hash_map};
@@ -21,18 +23,14 @@ use lsp::LanguageServerId;
 use parking_lot::{Mutex, RwLock};
 use postage::watch;
 
-use smallvec::SmallVec;
 use std::{
-    cell::LazyCell,
     ffi::OsStr,
-    ops::Not,
     path::{Path, PathBuf},
     sync::Arc,
 };
-use sum_tree::Bias;
-use text::{Point, Rope};
+use text::Rope;
 use theme::Theme;
-use unicase::UniCase;
+
 use util::{maybe, post_inc};
 
 pub struct LanguageRegistry {
@@ -46,7 +44,7 @@ struct LanguageRegistryState {
     next_language_server_id: usize,
     languages: Vec<Arc<Language>>,
     language_settings: AllLanguageSettingsContent,
-    available_languages: Vec<AvailableLanguage>,
+    available_languages: AvailableLanguages,
     grammars: HashMap<Arc<str>, AvailableGrammar>,
     lsp_adapters: HashMap<LanguageName, Vec<Arc<CachedLspAdapter>>>,
     all_lsp_adapters: HashMap<LanguageServerName, Arc<CachedLspAdapter>>,
@@ -68,40 +66,6 @@ pub struct FakeLanguageServerEntry {
     pub initializer: Option<Box<dyn 'static + Send + Sync + Fn(&mut lsp::FakeLanguageServer)>>,
     pub tx: futures::channel::mpsc::UnboundedSender<lsp::FakeLanguageServer>,
     pub _server: Option<lsp::FakeLanguageServer>,
-}
-
-#[derive(Clone)]
-pub struct AvailableLanguage {
-    id: LanguageId,
-    name: LanguageName,
-    grammar: Option<Arc<str>>,
-    matcher: LanguageMatcher,
-    hidden: bool,
-    load: Arc<dyn Fn() -> Result<LoadedLanguage> + 'static + Send + Sync>,
-    loaded: bool,
-    manifest_name: Option<ManifestName>,
-}
-
-impl AvailableLanguage {
-    pub fn name(&self) -> LanguageName {
-        self.name.clone()
-    }
-
-    pub fn matcher(&self) -> &LanguageMatcher {
-        &self.matcher
-    }
-
-    pub fn hidden(&self) -> bool {
-        self.hidden
-    }
-}
-
-#[derive(Copy, Clone, Default)]
-enum LanguageMatchPrecedence {
-    #[default]
-    Undetermined,
-    PathOrContent(usize),
-    UserConfigured(usize),
 }
 
 enum AvailableGrammar {
@@ -143,7 +107,7 @@ impl LanguageRegistry {
             state: RwLock::new(LanguageRegistryState {
                 next_language_server_id: 0,
                 languages: Vec::new(),
-                available_languages: Vec::new(),
+                available_languages: AvailableLanguages::default(),
                 grammars: Default::default(),
                 language_settings: Default::default(),
                 loading_languages: Default::default(),
@@ -414,26 +378,18 @@ impl LanguageRegistry {
     ) {
         let state = &mut *self.state.write();
 
-        for existing_language in &mut state.available_languages {
-            if existing_language.name == name {
-                existing_language.grammar = grammar_name;
-                existing_language.matcher = matcher;
-                existing_language.load = load;
-                existing_language.manifest_name = manifest_name;
-                return;
-            }
+        let was_added = state.available_languages.register(
+            name,
+            grammar_name,
+            matcher,
+            hidden,
+            manifest_name,
+            load,
+        );
+        if !was_added {
+            return;
         }
 
-        state.available_languages.push(AvailableLanguage {
-            id: LanguageId::new(),
-            name,
-            grammar: grammar_name,
-            matcher,
-            load,
-            hidden,
-            loaded: false,
-            manifest_name,
-        });
         state.version += 1;
         state.reload_count += 1;
         *state.subscription.0.borrow_mut() = ();
@@ -475,12 +431,13 @@ impl LanguageRegistry {
 
     pub fn language_names(&self) -> Vec<LanguageName> {
         let state = self.state.read();
-        let mut result = state
-            .available_languages
-            .iter()
-            .filter_map(|l| l.loaded.not().then_some(l.name.clone()))
-            .chain(state.languages.iter().map(|l| l.config.name.clone()))
-            .collect::<Vec<_>>();
+        let mut result = state.available_languages.unloaded_language_names();
+        result.extend(
+            state
+                .languages
+                .iter()
+                .map(|language| language.config.name.clone()),
+        );
         result.sort_unstable_by_key(|language_name| language_name.as_ref().to_lowercase());
         result
     }
@@ -495,7 +452,7 @@ impl LanguageRegistry {
     /// Add a pre-loaded language to the registry.
     pub fn add(&self, language: Arc<Language>) {
         let mut state = self.state.write();
-        state.available_languages.push(AvailableLanguage {
+        state.available_languages.add(AvailableLanguage {
             id: language.id,
             name: language.name(),
             grammar: language.config.grammar.clone(),
@@ -539,17 +496,8 @@ impl LanguageRegistry {
         self: &Arc<Self>,
         name: &str,
     ) -> impl Future<Output = Result<Arc<Language>>> + use<> {
-        let name = UniCase::new(name);
-        let rx = self.get_or_load_language(|language_name, _, current_best_match| {
-            match current_best_match {
-                LanguageMatchPrecedence::Undetermined if UniCase::new(&language_name.0) == name => {
-                    Some(LanguageMatchPrecedence::PathOrContent(name.len()))
-                }
-                LanguageMatchPrecedence::Undetermined
-                | LanguageMatchPrecedence::UserConfigured(_)
-                | LanguageMatchPrecedence::PathOrContent(_) => None,
-            }
-        });
+        let language = self.state.read().available_languages.find_by_name(name);
+        let rx = self.get_or_load_language(language);
         async move { rx.await? }
     }
 
@@ -557,12 +505,7 @@ impl LanguageRegistry {
         let available_language = {
             let state = self.state.read();
 
-            let Some(available_language) = state
-                .available_languages
-                .iter()
-                .find(|lang| lang.id == id)
-                .cloned()
-            else {
+            let Some(available_language) = state.available_languages.find_by_id(id) else {
                 anyhow::bail!(LanguageNotFound);
             };
             available_language
@@ -572,55 +515,27 @@ impl LanguageRegistry {
     }
 
     pub fn language_name_for_extension(self: &Arc<Self>, extension: &str) -> Option<LanguageName> {
-        self.state.try_read().and_then(|state| {
-            state
-                .available_languages
-                .iter()
-                .find(|language| {
-                    language
-                        .matcher()
-                        .path_suffixes
-                        .iter()
-                        .any(|suffix| *suffix == extension)
-                })
-                .map(|language| language.name.clone())
-        })
+        self.state
+            .try_read()
+            .and_then(|state| state.available_languages.find_name_by_extension(extension))
     }
 
     pub fn language_for_name_or_extension(
         self: &Arc<Self>,
         string: &str,
     ) -> impl Future<Output = Result<Arc<Language>>> {
-        let string = UniCase::new(string);
-        let rx = self.get_or_load_language(|name, config, current_best_match| {
-            let name_matches = || {
-                UniCase::new(&name.0) == string
-                    || config
-                        .path_suffixes
-                        .iter()
-                        .any(|suffix| UniCase::new(suffix) == string)
-            };
-
-            match current_best_match {
-                LanguageMatchPrecedence::Undetermined => {
-                    name_matches().then_some(LanguageMatchPrecedence::PathOrContent(string.len()))
-                }
-                LanguageMatchPrecedence::PathOrContent(len) => (string.len() > len
-                    && name_matches())
-                .then_some(LanguageMatchPrecedence::PathOrContent(string.len())),
-                LanguageMatchPrecedence::UserConfigured(_) => None,
-            }
-        });
+        let language = self
+            .state
+            .read()
+            .available_languages
+            .find_by_name_or_extension(string);
+        let rx = self.get_or_load_language(language);
         async move { rx.await? }
     }
 
     pub fn available_language_for_name(self: &Arc<Self>, name: &str) -> Option<AvailableLanguage> {
         let state = self.state.read();
-        state
-            .available_languages
-            .iter()
-            .find(|l| l.name.0.as_ref() == name)
-            .cloned()
+        state.available_languages.find_by_exact_name(name)
     }
 
     /// Look up a language by its modeline name (vim filetype or emacs mode).
@@ -633,32 +548,10 @@ impl LanguageRegistry {
         self: &Arc<Self>,
         modeline_name: &str,
     ) -> Option<AvailableLanguage> {
-        let modeline_name_lower = modeline_name.to_lowercase();
-        let state = self.state.read();
-
-        state
+        self.state
+            .read()
             .available_languages
-            .iter()
-            .find(|lang| {
-                lang.matcher
-                    .modeline_aliases
-                    .iter()
-                    .any(|alias| alias.to_lowercase() == modeline_name_lower)
-            })
-            .or_else(|| {
-                state.available_languages.iter().find(|lang| {
-                    lang.grammar
-                        .as_ref()
-                        .is_some_and(|g| g.to_lowercase() == modeline_name_lower)
-                })
-            })
-            .or_else(|| {
-                state
-                    .available_languages
-                    .iter()
-                    .find(|lang| lang.name.0.to_lowercase() == modeline_name_lower)
-            })
-            .cloned()
+            .find_by_modeline_name(modeline_name)
     }
 
     pub fn language_for_file(
@@ -703,164 +596,10 @@ impl LanguageRegistry {
         content: Option<&Rope>,
         user_file_types: Option<&FxHashMap<Arc<str>, (GlobSet, Vec<String>)>>,
     ) -> Option<AvailableLanguage> {
-        let filename = path.file_name().and_then(|filename| filename.to_str());
-        // `Path.extension()` returns None for files with a leading '.'
-        // and no other extension which is not the desired behavior here,
-        // as we want `.zshrc` to result in extension being `Some("zshrc")`
-        let extension = filename.and_then(|filename| filename.split('.').next_back());
-        let path_suffixes = [extension, filename, path.to_str()]
-            .iter()
-            .filter_map(|suffix| suffix.map(|suffix| (suffix, globset::Candidate::new(suffix))))
-            .collect::<SmallVec<[_; 3]>>();
-        let content = LazyCell::new(|| {
-            content.map(|content| {
-                let end = content.clip_point(Point::new(0, 256), Bias::Left);
-                let end = content.point_to_offset(end);
-                content.chunks_in_range(0..end).collect::<String>()
-            })
-        });
-        self.find_matching_language(move |language_name, config, current_best_match| {
-            let path_matches_default_suffix = || {
-                let len =
-                    config
-                        .path_suffixes
-                        .iter()
-                        .fold(0, |acc: usize, path_suffix: &String| {
-                            let ext = ".".to_string() + path_suffix;
-
-                            let matched_suffix_len = path_suffixes
-                                .iter()
-                                .find(|(suffix, _)| suffix.ends_with(&ext) || suffix == path_suffix)
-                                .map(|(suffix, _)| suffix.len());
-
-                            match matched_suffix_len {
-                                Some(len) => acc.max(len),
-                                None => acc,
-                            }
-                        });
-                (len > 0).then_some(len)
-            };
-
-            let path_matches_custom_suffix = || {
-                user_file_types
-                    .and_then(|types| types.get(language_name.as_ref()))
-                    .map_or(None, |(custom_suffixes, _)| {
-                        path_suffixes
-                            .iter()
-                            .find(|(_, candidate)| custom_suffixes.is_match_candidate(candidate))
-                            .map(|(suffix, _)| suffix.len())
-                    })
-            };
-
-            let content_matches = || {
-                config.first_line_pattern.as_ref().is_some_and(|pattern| {
-                    content
-                        .as_ref()
-                        .is_some_and(|content| pattern.is_match(content))
-                })
-            };
-
-            // Only return a match for the given file if we have a better match than
-            // the current one.
-            match current_best_match {
-                LanguageMatchPrecedence::PathOrContent(current_len) => {
-                    if let Some(len) = path_matches_custom_suffix() {
-                        // >= because user config should win tie with system ext len
-                        (len >= current_len).then_some(LanguageMatchPrecedence::UserConfigured(len))
-                    } else if let Some(len) = path_matches_default_suffix() {
-                        // >= because user config should win tie with system ext len
-                        (len >= current_len).then_some(LanguageMatchPrecedence::PathOrContent(len))
-                    } else {
-                        None
-                    }
-                }
-                LanguageMatchPrecedence::Undetermined => {
-                    if let Some(len) = path_matches_custom_suffix() {
-                        Some(LanguageMatchPrecedence::UserConfigured(len))
-                    } else if let Some(len) = path_matches_default_suffix() {
-                        Some(LanguageMatchPrecedence::PathOrContent(len))
-                    } else if content_matches() {
-                        Some(LanguageMatchPrecedence::PathOrContent(1))
-                    } else {
-                        None
-                    }
-                }
-                LanguageMatchPrecedence::UserConfigured(_) => None,
-            }
-        })
-    }
-
-    fn find_matching_language(
-        self: &Arc<Self>,
-        callback: impl Fn(
-            &LanguageName,
-            &LanguageMatcher,
-            LanguageMatchPrecedence,
-        ) -> Option<LanguageMatchPrecedence>,
-    ) -> Option<AvailableLanguage> {
-        let state = self.state.read();
-        let available_language = state
+        self.state
+            .read()
             .available_languages
-            .iter()
-            .rev()
-            .fold(None, |best_language_match, language| {
-                let current_match_type = best_language_match
-                    .as_ref()
-                    .map_or(LanguageMatchPrecedence::default(), |(_, score)| *score);
-                let language_score =
-                    callback(&language.name, &language.matcher, current_match_type);
-
-                match (language_score, current_match_type) {
-                    // no current best, so our candidate is better
-                    (
-                        Some(
-                            LanguageMatchPrecedence::PathOrContent(_)
-                            | LanguageMatchPrecedence::UserConfigured(_),
-                        ),
-                        LanguageMatchPrecedence::Undetermined,
-                    ) => language_score.map(|new_score| (language.clone(), new_score)),
-
-                    // our candidate is better only if the name is longer
-                    (
-                        Some(LanguageMatchPrecedence::PathOrContent(new_len)),
-                        LanguageMatchPrecedence::PathOrContent(current_len),
-                    )
-                    | (
-                        Some(LanguageMatchPrecedence::UserConfigured(new_len)),
-                        LanguageMatchPrecedence::UserConfigured(current_len),
-                    )
-                    | (
-                        Some(LanguageMatchPrecedence::PathOrContent(new_len)),
-                        LanguageMatchPrecedence::UserConfigured(current_len),
-                    ) => {
-                        if new_len > current_len {
-                            language_score.map(|new_score| (language.clone(), new_score))
-                        } else {
-                            best_language_match
-                        }
-                    }
-
-                    // our candidate is better if the name is longer or equal to
-                    (
-                        Some(LanguageMatchPrecedence::UserConfigured(new_len)),
-                        LanguageMatchPrecedence::PathOrContent(current_len),
-                    ) => {
-                        if new_len >= current_len {
-                            language_score.map(|new_score| (language.clone(), new_score))
-                        } else {
-                            best_language_match
-                        }
-                    }
-
-                    // no candidate, use current best
-                    (None, _) | (Some(LanguageMatchPrecedence::Undetermined), _) => {
-                        best_language_match
-                    }
-                }
-            })
-            .map(|(available_language, _)| available_language);
-        drop(state);
-        available_language
+            .find_for_file(path, content, user_file_types)
     }
 
     #[ztracing::instrument(skip_all)]
@@ -954,13 +693,9 @@ impl LanguageRegistry {
     #[ztracing::instrument(skip_all)]
     fn get_or_load_language(
         self: &Arc<Self>,
-        callback: impl Fn(
-            &LanguageName,
-            &LanguageMatcher,
-            LanguageMatchPrecedence,
-        ) -> Option<LanguageMatchPrecedence>,
+        language: Option<AvailableLanguage>,
     ) -> oneshot::Receiver<Result<Arc<Language>>> {
-        let Some(language) = self.find_matching_language(callback) else {
+        let Some(language) = language else {
             let (tx, rx) = oneshot::channel();
             let _ = tx.send(Err(anyhow!(LanguageNotFound)));
             return rx;
@@ -1141,9 +876,7 @@ impl LanguageRegistryState {
         self.languages.clear();
         self.version += 1;
         self.reload_count += 1;
-        for language in &mut self.available_languages {
-            language.loaded = false;
-        }
+        self.available_languages.mark_all_unloaded();
         *self.subscription.0.borrow_mut() = ();
     }
 
@@ -1188,8 +921,7 @@ impl LanguageRegistryState {
 
         self.languages
             .retain(|language| !languages_to_remove.contains(&language.name()));
-        self.available_languages
-            .retain(|language| !languages_to_remove.contains(&language.name));
+        self.available_languages.remove(languages_to_remove);
         self.grammars
             .retain(|name, _| !grammars_to_remove.contains(name));
         self.version += 1;
@@ -1200,12 +932,7 @@ impl LanguageRegistryState {
     /// Mark the given language as having been loaded, so that the
     /// language registry won't try to load it again.
     fn mark_language_loaded(&mut self, id: LanguageId) {
-        for language in &mut self.available_languages {
-            if language.id == id {
-                language.loaded = true;
-                break;
-            }
-        }
+        self.available_languages.mark_loaded(id);
     }
 }
